+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 ## HEAD (Unreleased)
 
+### Improvements
+
+- Expose options, parent, dependencies, and provider config (https://github.com/pulumi/pulumi-policy/pull/184).
+
 ## 0.4.0 (2020-01-30)
 
 ### Improvements
 
 - Add support for using `Config`, `getProject()`, `getStack()`, and `isDryRun()` from Policy Packs
-  via upgraded dependency on `@pulumi/pulumi` v1.8.0 (requires v1.8.0 of the Pulumi SDK).
+  via upgraded dependency on `@pulumi/pulumi` v1.8.0 (requires v1.8.0 or later of the Pulumi SDK) (https://github.com/pulumi/pulumi-policy/pull/169).
 
 - Provide easier type checking for `validateStack`, along with `isType` and `asType` helper functions
   (https://github.com/pulumi/pulumi-policy/pull/173).